task 6

+++ hello_world/6-size.c
@@ -10,14 +10,14 @@ int main(void)
 {
 char charType;
 int intType;
-int longType;
-int longLongType;
+long int longType;
+long long int longLongType;
 float floatType;
 
 printf("Size of char: %d byte(s)\n", sizeof(charType));
 printf("Size of int: %d byte(s)\n", sizeof(intType));
-printf("Size of long int: %ld byte(s)\n", sizeof(longType));
-printf("Size of long long int: %llu byte(s)\n", sizeof(longLongType));
-printf("Size of float: %ld byte(s)\n", sizeof(floatType));
+printf("Size of long int: %d byte(s)\n", sizeof(longType));
+printf("Size of long long int: %d byte(s)\n", sizeof(longLongType));
+printf("Size of float: %d byte(s)\n", sizeof(floatType));
 return (0);
 }
